Revert "Support macOS"
