+++ Example/Package.swift
@@ -1,11 +1,11 @@
-// swift-tools-version: 6.0
+// swift-tools-version: 5.9
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
 
 let package = Package(
     name: "PrefireExample",
-    platforms: [.iOS(.v17), .macOS(.v15)],
+    platforms: [.iOS("17")],
     products: [
         .library(
             name: "PrefireExample",
