+++ Example/PreFireExample.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,5 +1,5 @@
 {
-  "originHash" : "5ecf5b0090e0105b6cdfed96d506f54e9d024c0e213cca75273534ba9faa284d",
+  "originHash" : "ac988f1bdd201a27034cd4efa451ebe5806e0fd8b0ac226a19cbd79e07ebdb21",
   "pins" : [
     {
       "identity" : "accessibilitysnapshot",
@@ -33,8 +33,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/apple/swift-syntax.git",
       "state" : {
-        "revision" : "64889f0c732f210a935a0ad7cda38f77f876262d",
-        "version" : "509.1.1"
+        "revision" : "74203046135342e4a4a627476dd6caf8b28fe11b",
+        "version" : "509.0.0"
       }
     }
   ],
