+++ Example/Shared/Examples/UIKitView.swift
@@ -1,22 +1,14 @@
-import Prefire
 import SwiftUI
+import Prefire
 
 #Preview {
-    #if os(iOS)
     let view = UIView(frame: .init(origin: .zero, size: .init(width: 100, height: 100)))
     view.backgroundColor = .red
     return view
-    #else
-    EmptyView()
-    #endif
 }
 
 #Preview {
-    #if os(iOS)
     let viewController = UIViewController()
     viewController.view.backgroundColor = .green
     return viewController
-    #else
-    EmptyView()
-    #endif
 }
