+++ Templates/PreviewModels.stencil
@@ -4,9 +4,7 @@
 import SwiftUI
 import Prefire
 {% for import in argument.imports %}
-#if canImport({{ import }})
 import {{ import }}
-#endif
 {% endfor %}
 {% for import in argument.testableImports %}
 @testable import {{ import }}
